refactor 2.6.0 migration
