+++ src/internal/clusterstate/v2.5.0/clusterstate.go
@@ -37,18 +37,6 @@ func Migrate(state migrations.State) migrations.State {
 				return err
 			}
 			return nil
-		}).
-		Apply("Add commit_provenance table", func(ctx context.Context, env migrations.Env) error {
-			return pfsdb.SetupCommitProvenanceV0(ctx, env.Tx)
-		}).
-		Apply("Remove Alias Commits", func(ctx context.Context, env migrations.Env) error {
-			return migrateAliasCommits(ctx, env.Tx)
-		}).
-		Apply("Remove branch from the Commit key", func(ctx context.Context, env migrations.Env) error {
-			if err := migrateToBranchlessCommits(ctx, env.Tx); err != nil {
-				return err
-			}
-			return migrateBranchlessCommitsPPS(ctx, env.Tx)
 		})
 	// DO NOT MODIFY THIS STATE
 	// IT HAS ALREADY SHIPPED IN A RELEASE
