+++ src/internal/pfsdb/commit_provenance.go
@@ -140,6 +140,10 @@ func SetupCommitProvenanceV0(ctx context.Context, tx *pachsql.Tx) error {
 	_, err := tx.ExecContext(ctx, schema)
 	return errors.EnsureStack(err)
 }
+func SetupCommitProvenanceV01(ctx context.Context, tx *pachsql.Tx) error {
+	_, err := tx.ExecContext(ctx, schemaUpdate)
+	return errors.EnsureStack(err)
+}
 
 // TODO(acohen4): verify how postgres behaves when does altering a table affect foreign key constraints?
 var schema = `
@@ -147,11 +151,7 @@ var schema = `
 		int_id BIGSERIAL PRIMARY KEY,
 		commit_id VARCHAR(4096) UNIQUE,
                 commit_set_id VARCHAR(4096) NOT NULL,
-                created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-                CONSTRAINT fk_col_commit
-                  FOREIGN KEY(commit_id)
-                  REFERENCES collections.commits(key)
-                  ON DELETE CASCADE
+                created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 	);
 
 	CREATE TABLE pfs.commit_provenance (
@@ -172,4 +172,10 @@ var schema = `
 		to_id,
 		from_id
 	);
+`
+var schemaUpdate = `
+       ALTER TABLE pfs.commits ADD CONSTRAINT fk_col_commit
+                  FOREIGN KEY(commit_id)
+                  REFERENCES collections.commits(key)
+                  ON DELETE CASCADE
 `
