+++ src/server/pfs/server/commit_set.go
@@ -455,7 +455,7 @@ func traverseToEdges(startCommit *pfs.CommitInfo, skipSet map[string]*pfs.Commit
 // that commit set Y is subvenant to commit set X.
 // Now consider the same graph, with the addition of a commit s@Z that has q@Y in its subvenance.
 // In this case, CommitSetSubvenance(X) still evaluates to [p@Y]. But since a commit in 'Z', depends on a commit
-// in 'Y', we haven't yet computed all of 'X''s subvenant commit sets. Therefore,
+// in 'Y', we haven't yet computed all of 'X”s subvenant commit sets. Therefore,
 // we re-evaluate CommitSetSubvenance for each collected commit set until our resulting set becomes stable.
 func (d *driver) subvenantCommitSets(txnCtx *txncontext.TransactionContext, commitset *pfs.CommitSet) ([]*pfs.CommitSet, error) {
 	collectSubvCommitSets := func(setIDs map[string]struct{}) (map[string]struct{}, error) {
